Unify regression and classification models (they are actually the same...) for both resnet and pointnet
