--- config/model/classification_network/pointnet_fc.yaml
@@ -1,1 +1,0 @@
-_target_: watchmal.model.classifier.PointNetFullyConnected
