--- config/model/feature_extractor/pointnet.yaml
@@ -1,5 +1,0 @@
-_target_: watchmal.model.pointnet.PointNetFeat
-feature_transform: True
-k: 5
-mean_pool: 256
-num_output_channels: 256
