+++ config/model/gnn.yaml
@@ -1,4 +1,4 @@
 _target_: watchmal.model.gnn.GCN
 in_feat: 8
 h_feat: 8
-num_classes: 4
+num_output_channels: 4
