--- config/model/regression.yaml
@@ -1,2 +1,0 @@
-_target_: watchmal.model.regression.Regression
-_recursive_: false
