--- config/model/regression_alg/linear.yaml
@@ -1,2 +1,0 @@
-_target_: watchmal.model.regression.LinearRegression
-num_outputs: 3
