--- config/model/regression_alg/linear_energies_and_position.yaml
@@ -1,2 +1,0 @@
-_target_: watchmal.model.regression.LinearRegression
-num_outputs: 4
