--- config/model/regression_alg/resnet.yaml
@@ -1,2 +1,0 @@
-_target_: watchmal.model.regression.PassThrough
-num_outputs: 3
