+++ config/pointnet_train.yaml
@@ -3,9 +3,7 @@ dump_path: './outputs/'
 defaults:
     - data: iwcd_short
     - data/dataset: iwcd_short_pointnet
-    - model: classifier
-    - model/feature_extractor: pointnet
-    - model/classification_network: pointnet_fc
+    - model: pointnet
     - engine: classifier
     - tasks/train: train_pointnet
     - optimizers@tasks.train.optimizers: adam
