+++ config/resnet_test.yaml
@@ -6,9 +6,7 @@ dump_path: './outputs/'
 defaults:
     - data: iwcd_short
     - data/dataset: iwcd_cnn_short
-    - model: classifier
-    - model/feature_extractor: resnet18
-    - model/classification_network: resnet
+    - model: resnet18
     - engine: classifier
     - tasks/restore_state: restore_state
     - tasks/evaluate: test
