--- watchmal/model/metric.py
@@ -1,9 +1,0 @@
-import torch
-
-
-def accuracy(output, target):
-    with torch.no_grad():
-        pred = torch.argmax(output, dim=1)
-        assert pred.shape[0] == len(target)
-        correct = torch.sum(pred == target).item()
-    return correct / len(target)
